Rename useFilter hook to useFilterComponent [ci skip]
