+++ src/components/Filters/Pagination/CurrentPage.component.tsx
@@ -9,111 +9,127 @@ import { clamp } from "@/utils/numbers";
 import { StyledFilterPill } from "../components/FilterPill.component";
 
 const DirectionButton = styled.button`
-    height: 100%;
+  height: 100%;
 `;
 
 const InnerContainer = styled.div`
-	display: grid;
-	place-items: center;
+  display: grid;
+  place-items: center;
 
-    padding: 0 ${SIZE_XS};
+  padding: 0 ${SIZE_XS};
 `;
 
 const PageNumberContainer = styled.span`
-	color: ${(props) => props.theme.base.textColor};
-	background: ${(props) => props.theme.base.background};
+  color: ${(props) => props.theme.base.textColor};
+  background: ${(props) => props.theme.base.background};
 `;
 
 const PageNumber = styled.input`
-	font-size: ${FONT_SIZE_XS};
-	padding: 0 ${SIZE_XS};
+  font-size: ${FONT_SIZE_XS};
+  padding: 0 ${SIZE_XS};
 
-	border: none;
-	outline: none;
+  border: none;
+  outline: none;
 
-	color: ${(props) => props.theme.base.textColor};
-	background: ${(props) => props.theme.base.background};
+  color: ${(props) => props.theme.base.textColor};
+  background: ${(props) => props.theme.base.background};
 
-	&:focus {
-		outline: 1px solid ${(prop) => prop.theme.base.border};
-	}
+  &:focus {
+    outline: 1px solid ${(prop) => prop.theme.base.border};
+  }
 
-	width: 3rem;
-	border-bottom: 1px solid ${(prop) => prop.theme.base.border};
+  width: 3rem;
+  border-bottom: 1px solid ${(prop) => prop.theme.base.border};
 `;
 
 const PageCount = styled.div`
-    display: flex;
-    align-items: center;
+  display: flex;
+  align-items: center;
 
-    height: 100%;
-    width: max-content;
+  height: 100%;
+  width: max-content;
 
-	background: ${(props) => props.theme.base.background};
+  background: ${(props) => props.theme.base.background};
 `;
 
 const CurrentPage = React.forwardRef<HTMLInputElement, CurrentPageProps>(
-	({ currentPage, numPages, onSetPage }, ref) => {
-		const previousDisabled = currentPage === 0;
-		const nextDisabled = currentPage === numPages;
+  ({ currentPage, numPages, onSetPage }, ref) => {
+    const previousDisabled = currentPage === 0;
+    const nextDisabled = currentPage === numPages;
 
-		function setPageNumber(event: React.ChangeEvent<HTMLInputElement>) {
-			const value = Number.parseInt(event.target.value, 10);
-			if (Number.isNaN(value)) {
-				return;
-			}
+    function setPageNumber(event: React.ChangeEvent<HTMLInputElement>) {
+      const value = Number.parseInt(event.target.value, 10);
+      if (Number.isNaN(value)) {
+        return;
+      }
 
-			const newPageNumber = clamp(value - 1, 0, numPages);
-			onSetPage(newPageNumber);
-		}
-		return (
-			<StyledFilterPill style={{ backgroundColor: "white" }}>
-				<FillIn
-					disabled={previousDisabled}
-					borderRadiusCorners={{ topLeft: SIZE_MD, bottomLeft: SIZE_MD }}
-				>
-					<DirectionButton
-						aria-label="Previous Page"
-						disabled={previousDisabled}
-						onClick={() => onSetPage(currentPage - 1)}
-					>
-						<InnerContainer>
-							<PreviousIcon />
-						</InnerContainer>
-					</DirectionButton>
-				</FillIn>
-				<PageCount>
-					<InnerContainer>
-						<PageNumberContainer>
-							Page{" "}
-							<PageNumber
-								ref={ref}
-								aria-label="Set page number"
-								type="number"
-								value={currentPage + 1}
-								onChange={setPageNumber}
-							/>
-							/ {numPages + 1}
-						</PageNumberContainer>
-					</InnerContainer>
-				</PageCount>
-				<FillIn
-					disabled={nextDisabled}
-					borderRadiusCorners={{ topRight: SIZE_MD, bottomRight: SIZE_MD }}
-				>
-					<DirectionButton
-						aria-label="Next Page"
-						disabled={nextDisabled}
-						onClick={() => onSetPage(currentPage + 1)}
-					>
-						<InnerContainer>
-							<NextIcon />
-						</InnerContainer>
-					</DirectionButton>
-				</FillIn>
-			</StyledFilterPill>
-		);
-	},
+      const newPageNumber = clamp(value - 1, 0, numPages);
+      onSetPage(newPageNumber);
+    }
+
+    function handleKeydown(event: React.KeyboardEvent) {
+      if (event.key === "ArrowLeft") {
+        const newPageNumber = clamp(currentPage - 1, 0, numPages);
+        onSetPage(newPageNumber);
+        event.stopPropagation();
+      } else if (event.key === "ArrowRight") {
+        const newPageNumber = clamp(currentPage + 1, 0, numPages);
+        onSetPage(newPageNumber);
+        event.stopPropagation();
+      }
+    }
+
+    return (
+      <StyledFilterPill
+        onKeyDown={handleKeydown}
+        style={{ backgroundColor: "white" }}
+      >
+        <FillIn
+          disabled={previousDisabled}
+          borderRadiusCorners={{ topLeft: SIZE_MD, bottomLeft: SIZE_MD }}
+        >
+          <DirectionButton
+            aria-label="Previous Page"
+            disabled={previousDisabled}
+            onClick={() => onSetPage(currentPage - 1)}
+          >
+            <InnerContainer>
+              <PreviousIcon />
+            </InnerContainer>
+          </DirectionButton>
+        </FillIn>
+        <PageCount>
+          <InnerContainer>
+            <PageNumberContainer>
+              Page{" "}
+              <PageNumber
+                ref={ref}
+                aria-label="Set page number"
+                type="number"
+                value={currentPage + 1}
+                onChange={setPageNumber}
+              />
+              / {numPages + 1}
+            </PageNumberContainer>
+          </InnerContainer>
+        </PageCount>
+        <FillIn
+          disabled={nextDisabled}
+          borderRadiusCorners={{ topRight: SIZE_MD, bottomRight: SIZE_MD }}
+        >
+          <DirectionButton
+            aria-label="Next Page"
+            disabled={nextDisabled}
+            onClick={() => onSetPage(currentPage + 1)}
+          >
+            <InnerContainer>
+              <NextIcon />
+            </InnerContainer>
+          </DirectionButton>
+        </FillIn>
+      </StyledFilterPill>
+    );
+  }
 );
 
 export default CurrentPage;
